Integrating Environ Functionality
EnvironBox.addstack() now returns the environ number that the unit was
put into.
Each planet now has it's own environment (all currently the same)
Each environment now moves with it's planet.
Moving units gives an error, because the stack class does not have the
properties described in the original API for this class.