Units now 'stick' to environments
I modified Ranger's code to accept units instead of stacks for the
purposes of allowing units to stick with the environment they've been
dragged to after the planets have been moved/swapped about the screen.
This addition is by no means a finished product; in fact it's rather
hacky - it literally uses a 'bullshit' function to perform the task
needed. I have a more permanent solution which I will be implement later
this week.