The new and improved Dumb GUI
I re-wrote most of my code and made some major changes to Ranger's
environ code; after talking things over with Ranger we decided to make
the environs static. As the code currently stands, the values for the
star system, i.e. the planets and their environs are hard coded until we
have a way to prompt the back end for the needed data, and Joe's buttons
are not currently implemented, simply because I haven't gotten around to
it; the images have also all been scaled down so that the ships fit
nicely in the environ boxes, and to prevent the planets and environs
from clipping.